feat(linear_algebra/multilinear/basic): multilinear maps on basis vectors
Add two versions of the fact that a multilinear map on finitely many
arguments is determined by its values on basis vectors.  The precise
requirements for each version follow from the results used in the
proof: `basis.ext_multilinear_fin` uses the `curry` and `uncurry`
results to deduce it from `basis.ext` and thus works for multilinear
maps on a family of modules indexed by `fin n.succ`, while
`basis.ext_multilinear` works for an arbitrary `fintype` index type
but is limited to the case where the modules for all the arguments
(and the bases used for those modules) are the same, since it's
derived from the previous result using `dom_dom_congr` which only
handles the case where all the modules are the same.

This is in preparation for proving a corresponding result for
alternating maps, for which the case of all argument modules the same
suffices.

There is probably room for both golfing (the proofs seem longer than
should be necessary given the trivial content) and making more
general.

If the dependency of `linear_algebra.multilinear.basic` on
`linear_algebra.basis` isn't desired, this code could of course go in
a separate file.