chore(data/fin/basic): move tuple stuff to a new file
There are almost 600 lines of tuple stuff, which is definitely sufficient to justify a standalone file.

The author information has been guessed from the git history.
Floris is responsible for `cons` and `tail` which came first, Chris added find, and Yury and Sébastien were involved all over the place.