feat(linear_algebra/orientation): orientations of modules
Define orientations of modules, along the lines of a definition
suggested by @hrmacbeth: equivalence classes of nonzero alternating
maps.  See:
https://leanprover.zulipchat.com/#narrow/stream/217875-Is-there.20code.20for.20X.3F/topic/adding.20angles/near/243856522

The motivating use case is over the reals, but the definitions and
lemmas are for `ordered_comm_ring`, `linear_ordered_comm_ring` or
`linear_ordered_field` as appropriate (some might work for weaker type
classes, but a `nontrivial` `ordered_comm_ring` looks like it's the
weakest case for which much useful can be done with this definition).
The definition uses an arbitrary index type; the motivating use case
is where this has the cardinality of a basis (two-dimensional use
cases will use an index type that is definitionally `fin 2`, for
example).

Given an intended use case (oriented angles for Euclidean geometry)
where it will make sense for many proofs (and notation) to fix a
choice of orientation throughout, `orientation` is defined as a type
class so the choice of orientation can be implicit in such proofs and
the lemmas they use.  However, everything so far uses it as a
structure rather than having an orientation implicit through type
class inference; it's only for applications of orientations where it's
convenient to work with a fixed (but arbitrary) choice of orientation
that it will start to make sense to pass an orientation as a type
class argument.

I expect more definitions and lemmas about orientations will need
adding to make much use of orientations.  In particular, I expect to
need to add more about orientations in relation to bases
(e.g. extracting a basis that gives a given orientation, in positive
dimension).