feat(topology/connected.lean): add a theorem on connectedness of union
Prove that the Union of connected sets indexed by the vertices of a
graph is connected, when the graph is connected and any two sets
indexed by neighboring vertices meet.

Apply this to a new proof that the union of sets indexed by `nat`
is connected when a set and the next meet.

TODO: - apply this to sets indexed by an interval in `int`.