avoid explicit typing of spectrum
pass the scalar ring as an explicit type to the resolvent
and spectrum functions, and then include it in the notation
for the spectrum. This way Lean can properly find the type
of the spectrum (since A may be an algebra over various
rings). Also delete references to the explicit type of the
spectrum throughout.