feat(linear_algebra/affine_space/barycentric_coords): barycentric coordinates are ratio of determinants (#10320)
The main goal of this PR is the new lemma `affine_basis.det_smul_coords_eq_camer_coords`

A secondary goal is a minor refactor of barycentric coordinates so that they are associated to a new structure `affine_basis`. As well as making the API for affine spaces more similar to that of modules, this enables an extremely useful dot notation.

The work here could easily be split into two PRs and I will happily do so if requested.

Formalized as part of the Sphere Eversion project.