feat(analyis/normed_space/spectrum and algebra/algebra/spectrum): prove properties of spectrum in a Banach algebra
Prove that the `spectrum` of an element in a Banach algebra
is closed and bounded, hence compact when the scalar field is
proper. Compute the derivative of the `resolvent a` function
in preparation for showing that the spectrum is compact when
the base field is ℂ. Define the `spectral_radius` and prove
that it is bounded by the norm.